Fix firebase.json so it does not ignore fonts
firebase.json file attempts to provide a whitelist semantic in its "ignores" property for firebase hosting, so the only assets that are uploaded are the ones in dist/_next/static. This prevents, however, fonts from being uploaded (since they reside in dist/fonts after upload). The "ignores" list, unfortunately, treats every entry as an exclusion set, and uses the joint set of all these to obtain the complete set of ignored files. This does not allow overriding of ignore rules with ones with more specificity, so there is no way of stating "ignore everything in this folder, except this". Turning the list into a blacklist, and stating the folders and files that need to be ignored, instead of specifying the ones which do not, fixes this problem.

Fixes #40.

PR: #41.